% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/academic_year.R
\name{academic_year}
\alias{academic_year}
\title{Maps a date to an academic year.}
\usage{
academic_year(x)
}
\arguments{
\item{x}{A date to map to an acedemic year.}
}
\value{
An academic year as a charachter.
}
\description{
Maps a date to an academic year.
}
\examples{
academic_year(as.Date('2017-07-31'))
academic_year(as.Date('2017-08-01'))
}
