% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_uun.R
\name{clean_uun}
\alias{clean_uun}
\title{Produces a 'clean' uun.}
\usage{
clean_uun(uun_list)
}
\arguments{
\item{x}{A vector of uun's to be cleaned.}
}
\value{
A vector of clean uun's.
}
\description{
This function converts UUN's to a clean format. This is a string beginning with an upper case 'S' and followed by 7 digits.
}
\examples{
clean_uun('s1234567')
}
