% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_academic_year.R
\name{fix_academic_year}
\alias{fix_academic_year}
\title{Converts academic year into the correct format}
\usage{
fix_academic_year(x)
}
\arguments{
\item{x}{A vector of academic years in 'yyyy/y' format.}
}
\value{
A vector of academic year's in the corrected 'yyyy/yy' format.
}
\description{
Converts academic year into the correct format
}
\examples{
fix_academic_year('2014/5')
fix_academic_year(c('2014/5', '1999/0'))
fix_academic_year('2014/65')
}
