% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_read_excel_sheets.R
\name{lb_read_excel_sheets}
\alias{lb_read_excel_sheets}
\title{Read multiple sheets from an excel workbook}
\usage{
lb_read_excel_sheets(x, ignore = NULL, bind = TRUE, ...)
}
\arguments{
\item{x}{An excel workbook.}

\item{ignore}{A vector of sheet names to ignore.}

\item{bind}{If TRUE then return a single data frame with id column 'Sheet'. Otherwise returns a list of data frames.}

\item{...}{Additional arguments to be passed to read_excel.}
}
\value{
Reads all sheets from an excel file into R as a single data frame, or if bind = FALSE a list of data frames.
}
\description{
Read multiple sheets from an excel workbook
}
\examples{
# Download Guardian university league table spreadsheet and save to desktop.
# https://docs.google.com/spreadsheets/d/1P1AiFGB5UFO0v26Ufe_3YNOT4lISK6DAaC8dvDRiVzs/edit#gid=1544561606

lb_read_excel_sheets(x = "~/Desktop/Guardian University Guide 2018.xlsx",
  ignore = c('Institutional', 'Sheet2', 'Sheet3'),
  skip = 1,
  n_max = 159)
}
