% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_calendar_years.R
\name{non-calendar}
\alias{non-calendar}
\alias{academic_year}
\alias{financial_year}
\title{Maps a date to a non-calendar year.}
\usage{
academic_year(x)

financial_year(x)
}
\arguments{
\item{x}{A date to map to an non-calendar year.}
}
\value{
An non-calendar year as a charachter.
}
\description{
Maps a date to a non-calendar year.
}
\note{
Academic years run from August 1st to July 31st.
 Financial years run from April 1st to March 31st.
}
\examples{
academic_year(as.Date('2017-07-31'))
academic_year(as.Date('2017-08-01'))

financial_year(as.Date('2017-03-31'))
financial_year(as.Date('2017-04-01'))

}
