% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{percent}
\alias{percent}
\title{Convert a proportion to a percentage.}
\usage{
percent(x, digits = 1)
}
\arguments{
\item{x}{A value to convert to a percentage.}

\item{digits}{The number of decimal places to round to, passed to the round function.}
}
\value{
percent simply takes a proprtion and converts to a percentage by multiplying by 100 and rounding.
}
\description{
Convert a proportion to a percentage.
}
\examples{
percent(0.43569)
percent(0.43569, digits = 2)
}
