% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_na.R
\name{replace_na}
\alias{replace_na}
\title{Replace NA values with a string.}
\usage{
replace_na(x, string)
}
\arguments{
\item{x}{A vector.}

\item{string}{A string}
}
\value{
A character vector.
}
\description{
Replace NA values with a string.
}
\examples{
replace_na(NA, 'Missing')
}
