% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankeymatic.R
\name{sankeymatic}
\alias{sankeymatic}
\title{Outputs text for pasting into sankeymatic.com}
\usage{
sankeymatic(df, left, right)
}
\arguments{
\item{df}{A data frame from which to creat the sankey chart.}

\item{left}{A variable from the data frame that will feature on the left hand side of the sankey chart.}

\item{right}{A variable from the data frame that will feature on the right hand side of the sankey chart.}
}
\value{
A data.frame with one column, 'sankeymatic', of nodes and counts in the format required by sankeymatic.com.

A data.frame is produced rather than a tibble as the print method allows row.names and right align to be suppressed. Text can then be copied directly from the console.
}
\description{
This function produces a data.frame of text that can be pasted into sankeymatic.com. It is essentially a stop gap function until some nice way of producing sankey charts in ggplot2 is introduced.
}
\examples{
library(ggplot2)
sankeymatic(diamonds, cut, color)
sankeymatic(diamonds, cut, color) \%>\%  print(right = FALSE, row.names = FALSE)
}
