% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_total.R
\name{summarise_total}
\alias{summarise_total}
\title{Summarise with totals}
\usage{
summarise_total(df, label = "Total", ...)
}
\arguments{
\item{df}{A grouped date.frame.}

\item{label}{The label to use in the grouping columns of the 'Total' row. Defaults to 'Total'.}

\item{...}{Name-value pairs of summary functions, as for dplyr:summarise. The name will be the name of the variable in the result. The value should be an expression that returns a single value like min(x), n(), or sum(is.na(y)).}
}
\value{
A data.frame containing a row of summary statistics for each group, plus a 'Total' row.
}
\description{
A version of dplyr's summarise function that provides an additional total row.
}
\examples{
iris \%>\% group_by(Species) \%>\% summarise(n = n(), `Mean Sepal Length` = mean(Sepal.Length))
iris \%>\% group_by(Species) \%>\% summarise_total(n = n(), `Mean Sepal Length` = mean(Sepal.Length))

}
\seealso{
summarise
}
