% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disclosure_control.R
\name{disclosure_control}
\alias{disclosure_control}
\title{Apply homelessness disclosure control}
\usage{
disclosure_control(x)
}
\arguments{
\item{x}{A numeric vector or a data.frame.}
}
\value{
A numeric vector or a data.frame.
}
\description{
This function applies a standard method of disclosure control used for example in the Scottish Government.
If x is a data.frame then rounding is applied to all numeric variables.
}
\examples{
disclosure_control(0:10)
disclosure_control(iris)
}
