% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_prop.R
\name{lb_prop}
\alias{lb_prop}
\title{Calculate a proportion from a vector.}
\usage{
lb_prop(x, numerator, denominator, percent = TRUE, digits = 1)
}
\arguments{
\item{x}{A vector to calculate a proportion from.}

\item{numerator}{A vector of elements to count in the numerator.}

\item{denominator}{A vector of elements to count in the denominator in addition to those included in the numerator.}

\item{digits}{The number of digits to round to if converting proportion to percentage.}
}
\value{
The proportion of specified values, presented as a percentage or proportion.
}
\description{
Calculate a proportion from a vector.
}
\examples{
lb_prop(x = c(T, T, T, T, F, NA), numerator = T, denominator = F)
lb_prop(x = c(T, T, T, T, F, NA), numerator = T, denominator = F, percent = FALSE)
}
