% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_n.R
\name{round_n}
\alias{round_n}
\title{Round to a specified integer}
\usage{
round_n(x, n)
}
\arguments{
\item{x}{A numeric vector.}

\item{n}{An integer to round to.}
}
\value{
A numeric vector.
}
\description{
This function rounds a number to a specified integer.
}
\examples{
round_n(16, 5)
round_n(c(7, 16, NA), 5)
}
